filtertype
