+++ modules/service/src/main/scala/lucuma/odb/graphql/snippet/table/FilterTypeTable.scala
@@ -1,0 +1,20 @@
+// Copyright (c) 2016-2022 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.graphql
+package snippet
+package table
+
+import edu.gemini.grackle.skunk.SkunkMapping
+import lucuma.odb.util.Codecs._
+import skunk.codec.all._
+
+trait FilterTypeTable[F[_]] { self: SkunkMapping[F] =>
+
+  object FilterTypeTable extends TableDef("t_filter_type") {
+    val Tag       = col("c_tag", tag)
+    val ShortName = col("c_short_name", varchar)
+    val LongName  = col("c_long_name", varchar)
+  }
+
+}
